Extract electron-builder raw config from build scripts
